+++ src/swmr.rs
@@ -1,15 +1,17 @@
-mod c;
+mod reader;
 mod wal;
+mod writer;
 
 /// An ordered write-ahead log implementation for multiple threads environments.
 pub mod base {
   use dbutils::checksum::Crc32;
 
-  use crate::pointer::Pointer;
+  use super::{reader, writer};
 
-  use super::wal;
-
-  pub use crate::base::{Reader, Writer};
+  pub use crate::wal::bytes::{
+    base::{Reader, Writer},
+    pointer::Pointer,
+  };
 
   /// An ordered write-ahead log implementation for multiple threads environments.
   ///
@@ -28,18 +30,22 @@ pub mod base {
   /// |         ...          |            ...          |         ...        |          ...        |        ...      |         ...        |
   /// +----------------------+-------------------------+--------------------+---------------------+-----------------+--------------------+
   /// ```
-  pub type OrderWal<C, S = Crc32> = wal::OrderWal<Pointer<C>, C, S>;
+  pub type OrderWal<C, S = Crc32> = writer::OrderWal<Pointer<C>, C, S>;
+
+  /// Immutable reader for the ordered write-ahead log [`OrderWal`].
+  pub type OrderWalReader<C, S = Crc32> = reader::OrderWalReader<Pointer<C>, C, S>;
 }
 
 /// A multiple version ordered write-ahead log implementation for multiple threads environments.
 pub mod multiple_version {
   use dbutils::checksum::Crc32;
 
-  use crate::pointer::VersionPointer;
+  use super::{reader, writer};
 
-  use super::wal;
-
-  pub use crate::mvcc::{Reader, Writer};
+  pub use crate::wal::bytes::{
+    mvcc::{Reader, Writer},
+    pointer::VersionPointer,
+  };
 
   /// A multiple versioned ordered write-ahead log implementation for multiple threads environments.
   ///
@@ -56,18 +62,22 @@ pub mod multiple_version {
   /// |         ...          |            ...          |         ...        |          ...        |        ...          |         ...     |        ,,,         |
   /// +----------------------+-------------------------+--------------------+---------------------+---------------------+-----------------+--------------------+
   /// ```
-  pub type OrderWal<C, S = Crc32> = wal::OrderWal<VersionPointer<C>, C, S>;
+  pub type OrderWal<C, S = Crc32> = writer::OrderWal<VersionPointer<C>, C, S>;
+
+  /// Immutable reader for the multiple versioned ordered write-ahead log [`OrderWal`].
+  pub type OrderWalReader<C, S = Crc32> = reader::OrderWalReader<VersionPointer<C>, C, S>;
 }
 
 /// The ordered write-ahead log only supports generic.
 pub mod generic {
   use dbutils::checksum::Crc32;
 
-  use crate::generic::GenericPointer;
+  use super::{reader, writer};
 
-  use super::wal;
-
-  pub use crate::generic::base::{Reader, Writer};
+  pub use crate::wal::generic::{
+    base::{Reader, Writer},
+    GenericPointer,
+  };
 
   /// A generic ordered write-ahead log implementation for multiple threads environments.
   ///
@@ -86,18 +96,24 @@ pub mod generic {
   /// |         ...          |            ...          |         ...        |          ...        |        ...      |         ...        |
   /// +----------------------+-------------------------+--------------------+---------------------+-----------------+--------------------+
   /// ```
-  pub type GenericOrderWal<K, V, S = Crc32> = wal::GenericOrderWal<K, V, GenericPointer<K, V>, S>;
+  pub type GenericOrderWal<K, V, S = Crc32> =
+    writer::GenericOrderWal<K, V, GenericPointer<K, V>, S>;
+
+  /// Immutable reader for the generic ordered write-ahead log [`GenericOrderWal`].
+  pub type GenericOrderWalReader<K, V, S = Crc32> =
+    reader::GenericOrderWalReader<K, V, GenericPointer<K, V>, S>;
 }
 
 /// A multiple version ordered write-ahead log implementation for multiple threads environments.
 pub mod generic_multiple_version {
   use dbutils::checksum::Crc32;
 
-  use crate::generic::GenericVersionPointer;
+  use super::{reader, writer};
 
-  use super::wal;
-
-  pub use crate::generic::mvcc::{Reader, Writer};
+  pub use crate::wal::generic::{
+    mvcc::{Reader, Writer},
+    GenericVersionPointer,
+  };
 
   /// A multiple versioned generic ordered write-ahead log implementation for multiple threads environments.
   ///
@@ -115,5 +131,9 @@ pub mod generic_multiple_version {
   /// +----------------------+-------------------------+--------------------+---------------------+---------------------+-----------------+--------------------+
   /// ```
   pub type GenericOrderWal<K, V, S = Crc32> =
-    wal::GenericOrderWal<K, V, GenericVersionPointer<K, V>, S>;
+    writer::GenericOrderWal<K, V, GenericVersionPointer<K, V>, S>;
+
+  /// Immutable reader for the multiple versioned generic ordered write-ahead log [`GenericOrderWal`].
+  pub type GenericOrderWalReader<K, V, S = Crc32> =
+    reader::GenericOrderWalReader<K, V, GenericVersionPointer<K, V>, S>;
 }
