+++ src/swmr/wal.rs
@@ -1,215 +1,234 @@
-use crate::{
-  generic::{GenericComparator, GenericPointer, GenericVersionPointer},
-  pointer::{Pointer, VersionPointer},
-  sealed::{self, Constructable},
-  Ascend,
-};
-use dbutils::{checksum::Crc32, traits::Type, Comparator};
-use rarena_allocator::Allocator;
+use core::ops::{Bound, RangeBounds};
 
-// pub use crate::{
-//   batch::{Batch, BatchWithBuilders, BatchWithKeyBuilder, BatchWithValueBuilder},
-//   builder::Builder,
-//   wal::{Reader, Wal},
-//   Comparator, KeyBuilder, VacantBuffer, ValueBuilder,
-// };
-
-use core::{cell::UnsafeCell, marker::PhantomData};
+use crossbeam_skiplist::SkipSet;
+use dbutils::equivalent::Comparable;
 use rarena_allocator::sync::Arena;
-use std::sync::Arc;
 
-mod reader;
-pub use reader::*;
+use crate::{
+  sealed::{self, Pointer, Wal},
+  Options,
+};
 
-use super::c::OrderCore;
+impl<P> sealed::Memtable for SkipSet<P>
+where
+  P: Send + Ord,
+{
+  type Pointer = P;
 
-#[cfg(all(
-  test,
-  any(
-    all_tests,
-    test_swmr_constructor,
-    test_swmr_insert,
-    test_swmr_get,
-    test_swmr_iters,
-  )
-))]
-mod tests;
+  type Item<'a>
+    = crossbeam_skiplist::set::Entry<'a, P>
+  where
+    Self::Pointer: 'a,
+    Self: 'a;
 
-/// A ordered write-ahead log implementation for concurrent thread environments.
-pub struct GenericOrderWal<K: ?Sized, V: ?Sized, P, S = Crc32> {
-  core: Arc<UnsafeCell<OrderCore<P, GenericComparator<K>, S>>>,
-  _s: PhantomData<S>,
-  _v: PhantomData<V>,
+  type Iterator<'a>
+    = crossbeam_skiplist::set::Iter<'a, P>
+  where
+    Self::Pointer: 'a,
+    Self: 'a;
+
+  type Range<'a, Q, R>
+    = crossbeam_skiplist::set::Range<'a, Q, R, Self::Pointer>
+  where
+    Self::Pointer: 'a,
+    Self: 'a,
+    R: RangeBounds<Q>,
+    Q: Ord + ?Sized + Comparable<Self::Pointer>;
+
+  fn insert(&mut self, ele: Self::Pointer)
+  where
+    P: Ord + 'static,
+  {
+    SkipSet::insert(self, ele);
+  }
+
+  #[inline]
+  fn first(&self) -> Option<Self::Item<'_>> {
+    SkipSet::front(self)
+  }
+
+  #[inline]
+  fn last(&self) -> Option<Self::Item<'_>> {
+    SkipSet::back(self)
+  }
+
+  #[inline]
+  fn get<Q>(&self, key: &Q) -> Option<Self::Item<'_>>
+  where
+    Q: Ord + ?Sized + Comparable<P>,
+  {
+    SkipSet::get(self, key)
+  }
+
+  #[inline]
+  fn contains<Q>(&self, key: &Q) -> bool
+  where
+    Q: Ord + ?Sized + Comparable<P>,
+  {
+    SkipSet::contains(self, key)
+  }
+
+  #[inline]
+  fn iter(&self) -> Self::Iterator<'_> {
+    SkipSet::iter(self)
+  }
+
+  #[inline]
+  fn range<Q, R>(&self, range: R) -> Self::Range<'_, Q, R>
+  where
+    R: RangeBounds<Q>,
+    Q: Ord + ?Sized + Comparable<P>,
+  {
+    SkipSet::range(self, range)
+  }
 }
 
-impl<K, V, P, S> Constructable for GenericOrderWal<K, V, P, S>
+pub struct OrderCore<P, C, S> {
+  pub(super) arena: Arena,
+  pub(super) map: SkipSet<P>,
+  pub(super) max_version: u64,
+  pub(super) min_version: u64,
+  pub(super) opts: Options,
+  pub(super) cmp: C,
+  pub(super) cks: S,
+}
+
+impl<P, C, S> Wal<P, C, S> for OrderCore<P, C, S>
 where
-  K: ?Sized + 'static,
-  V: ?Sized + 'static,
-  S: 'static,
-  P: sealed::Pointer<Comparator = GenericComparator<K>> + Ord + Send + 'static,
+  P: Ord + Send + 'static,
 {
   type Allocator = Arena;
-  type Core = OrderCore<Self::Pointer, Self::Comparator, Self::Checksumer>;
-  type Pointer = P;
-  type Checksumer = S;
-  type Comparator = GenericComparator<K>;
-  type Reader = GenericOrderWalReader<K, V, P, S>;
+  type Memtable = SkipSet<P>;
 
   #[inline]
-  fn as_core(&self) -> &Self::Core {
-    unsafe { &*self.core.get() }
+  fn memtable(&self) -> &Self::Memtable {
+    &self.map
   }
 
   #[inline]
-  fn as_core_mut(&mut self) -> &mut Self::Core {
-    unsafe { &mut *self.core.get() }
-  }
-
-  #[inline]
-  fn from_core(core: Self::Core) -> Self {
+  fn construct(
+    arena: Arena,
+    set: SkipSet<P>,
+    opts: Options,
+    cmp: C,
+    checksumer: S,
+    maximum_version: u64,
+    minimum_version: u64,
+  ) -> Self {
     Self {
-      core: Arc::new(UnsafeCell::new(core)),
-      _s: PhantomData,
-      _v: PhantomData,
+      arena,
+      map: set,
+      cmp,
+      opts,
+      max_version: maximum_version,
+      min_version: minimum_version,
+      cks: checksumer,
     }
   }
-}
-
-impl<K, V, P, S> GenericOrderWal<K, V, P, S>
-where
-  K: ?Sized + 'static,
-  V: ?Sized + 'static,
-  S: 'static,
-  P: sealed::Pointer<Comparator = GenericComparator<K>> + Ord + Send + 'static,
-{
-  /// Returns the path of the WAL if it is backed by a file.
-  ///
-  /// ## Example
-  ///
-  /// ```rust
-  /// use orderwal::{unsync::GenericOrderWal, Wal, Builder};
-  ///
-  /// // A in-memory WAL
-  /// let wal = Builder::new().with_capacity(100).alloc::<GenericOrderWal>().unwrap();
-  ///
-  /// assert!(wal.path_buf().is_none());
-  /// ```
-  pub fn path_buf(&self) -> Option<&std::sync::Arc<std::path::PathBuf>> {
-    self.as_core().arena.path()
-  }
-}
-
-impl<K, V, S> crate::generic::base::Writer<K, V> for GenericOrderWal<K, V, GenericPointer<K, V>, S>
-where
-  K: ?Sized + Type + Ord + 'static,
-  V: ?Sized + Type + 'static,
-  GenericPointer<K, V>: Ord,
-  S: 'static,
-{
-  // type Reader = GenericOrderWalReader<K, V, GenericPointer<K, V>, S>;
 
   #[inline]
-  fn reader(&self) -> Self::Reader {
-    GenericOrderWalReader::new(self.core.clone())
+  fn options(&self) -> &Options {
+    &self.opts
   }
-}
 
-impl<K, V, S> crate::generic::mvcc::Writer<K, V>
-  for GenericOrderWal<K, V, GenericVersionPointer<K, V>, S>
-where
-  K: ?Sized + Type + Ord + 'static,
-  V: ?Sized + Type + 'static,
-  GenericVersionPointer<K, V>: Ord,
-  S: 'static,
-{
-  // type Reader = GenericOrderWalReader<K, V, GenericVersionPointer<K, V>, S>;
-
+  /// Returns the number of entries in the WAL.
   #[inline]
-  fn reader(&self) -> Self::Reader {
-    GenericOrderWalReader::new(self.core.clone())
+  fn len(&self) -> usize {
+    self.map.len()
   }
-}
 
-/// An ordered write-ahead log implementation for single thread environments.
-pub struct OrderWal<P, C = Ascend, S = Crc32> {
-  core: Arc<UnsafeCell<OrderCore<P, C, S>>>,
-  _s: PhantomData<S>,
-}
-
-impl<P, C, S> Constructable for OrderWal<P, C, S>
-where
-  C: 'static,
-  S: 'static,
-  P: sealed::Pointer<Comparator = C> + Ord + Send + 'static,
-{
-  type Allocator = Arena;
-  type Core = OrderCore<P, C, S>;
-  type Pointer = P;
-  type Checksumer = S;
-  type Comparator = C;
-  type Reader = OrderWalReader<P, C, S>;
-
+  /// Returns `true` if the WAL is empty.
   #[inline]
-  fn as_core(&self) -> &Self::Core {
-    unsafe { &*self.core.get() }
+  fn is_empty(&self) -> bool {
+    self.map.is_empty()
   }
 
   #[inline]
-  fn as_core_mut(&mut self) -> &mut Self::Core {
-    unsafe { &mut *self.core.get() }
+  fn maximum_version(&self) -> u64 {
+    self.max_version
   }
 
   #[inline]
-  fn from_core(core: Self::Core) -> Self {
-    Self {
-      core: Arc::new(UnsafeCell::new(core)),
-      _s: PhantomData,
+  fn minimum_version(&self) -> u64 {
+    self.min_version
+  }
+
+  #[inline]
+  fn update_maximum_version(&mut self, version: u64) {
+    self.max_version = version;
+  }
+
+  #[inline]
+  fn update_minimum_version(&mut self, version: u64) {
+    self.min_version = version;
+  }
+
+  #[inline]
+  fn allocator(&self) -> &Self::Allocator {
+    &self.arena
+  }
+
+  #[inline]
+  fn upper_bound<Q>(&self, version: Option<u64>, bound: Bound<&Q>) -> Option<&[u8]>
+  where
+    P: Pointer<Comparator = C>,
+    Q: Ord + ?Sized + Comparable<P>,
+  {
+    match version {
+      None => self.map.upper_bound(bound).map(|ent| ent.as_key_slice()),
+      Some(version) => {
+        let mut ent = self.map.upper_bound(bound);
+        loop {
+          match ent {
+            Some(ent) if ent.version() <= version => return Some(ent.as_key_slice()),
+            Some(e) => ent = e.next(),
+            None => return None,
+          }
+        }
+      }
     }
   }
-}
 
-impl<P, C, S> OrderWal<P, C, S>
-where
-  P: sealed::Pointer<Comparator = C> + Ord + Send + 'static,
-  C: 'static,
-  S: 'static,
-{
-  /// Returns the path of the WAL if it is backed by a file.
-  ///
-  /// ## Example
-  ///
-  /// ```rust
-  /// use orderwal::{unsync::OrderWal, Wal, Builder};
-  ///
-  /// // A in-memory WAL
-  /// let wal = Builder::new().with_capacity(100).alloc::<OrderWal>().unwrap();
-  ///
-  /// assert!(wal.path_buf().is_none());
-  /// ```
-  pub fn path_buf(&self) -> Option<&std::sync::Arc<std::path::PathBuf>> {
-    self.as_core().arena.path()
-  }
-}
-
-impl<C, S> crate::base::Writer for OrderWal<Pointer<C>, C, S>
-where
-  C: Comparator + Send + 'static,
-  S: 'static,
-{
   #[inline]
-  fn reader(&self) -> Self::Reader {
-    OrderWalReader::new(self.core.clone())
+  fn lower_bound<Q>(&self, version: Option<u64>, bound: core::ops::Bound<&Q>) -> Option<&[u8]>
+  where
+    P: Pointer<Comparator = C>,
+    Q: Ord + ?Sized + Comparable<P>,
+  {
+    match version {
+      None => self.map.lower_bound(bound).map(|ent| ent.as_key_slice()),
+      Some(version) => {
+        let mut ent = self.map.lower_bound(bound);
+        loop {
+          match ent {
+            Some(ent) if ent.version() <= version => return Some(ent.as_key_slice()),
+            Some(e) => ent = e.next(),
+            None => return None,
+          }
+        }
+      }
+    }
   }
-}
 
-impl<C, S> crate::mvcc::Writer for OrderWal<VersionPointer<C>, C, S>
-where
-  C: Comparator + Send + 'static,
-  S: 'static,
-{
   #[inline]
-  fn reader(&self) -> Self::Reader {
-    OrderWalReader::new(self.core.clone())
+  fn hasher(&self) -> &S {
+    &self.cks
+  }
+
+  #[inline]
+  fn comparator(&self) -> &C {
+    &self.cmp
+  }
+
+  #[inline]
+  fn insert_pointer(&mut self, ptr: P) {
+    self.map.insert(ptr);
+  }
+
+  #[inline]
+  fn insert_pointers(&mut self, ptrs: impl Iterator<Item = P>) {
+    for ptr in ptrs {
+      self.map.insert(ptr);
+    }
   }
 }
