--- src/unsync/tests.rs
@@ -1,15 +1,0 @@
-use super::*;
-
-#[cfg(all(test, any(test_unsync_constructor, all_tests)))]
-mod constructor;
-
-#[cfg(all(test, any(test_unsync_insert, all_tests)))]
-mod insert;
-
-#[cfg(all(test, any(test_unsync_iters, all_tests)))]
-mod iter;
-
-#[cfg(all(test, any(test_unsync_get, all_tests)))]
-mod get;
-
-const MB: u32 = 1024 * 1024;
