--- src/unsync/tests/constructor.rs
@@ -1,3 +1,0 @@
-use super::*;
-
-common_unittests!(unsync::constructor::OrderWal);
