--- src/unsync/tests/get.rs
@@ -1,3 +1,0 @@
-use super::*;
-
-common_unittests!(unsync::get::OrderWal);
