--- src/unsync/tests/insert.rs
@@ -1,5 +1,0 @@
-use super::*;
-
-common_unittests!(unsync::insert::OrderWal);
-
-common_unittests!(unsync::insert_batch::OrderWal);
