--- src/unsync/tests/iter.rs
@@ -1,3 +1,0 @@
-use super::*;
-
-common_unittests!(unsync::iters::OrderWal);
