+++ src/wal/generic/entry.rs
