+++ src/wal/generic/pointer.rs
