+++ src/wal/generic/query.rs
